arthmitra/backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
sqlalchemy
psycopg2-binary
pdfplumber
numpy
scipy
python-multipart
httpx
alembic
bcrypt>=4.0
python-jose[cryptography]
email-validator

